fix: numerous fixes for 2.4.1 update
see PR #2865 for more info
