+++ docusaurus.config.js
@@ -175,7 +175,6 @@ module.exports = {
     },
     colorMode: {
       defaultMode: "light",
-      disableSwitch: false,
       respectPrefersColorScheme: true,
     },
     prism: {
