+++ src/css/custom.css
@@ -8,6 +8,7 @@
 /* You can override the default Infima variables here. */
 :root {
   --ifm-color-primary: #3162ac;
+  --docusaurus-details-decoration-color: #3162ac !important;
   --ifm-color-primary-dark: #2c589b;
   --ifm-color-primary-darker: #2a5392;
   --ifm-color-primary-darkest: #224578;
@@ -25,6 +26,7 @@
 
 html[data-theme="dark"] {
   --ifm-color-primary: #5192c7;
+  --docusaurus-details-decoration-color: #5192c7 !important;
   --ifm-link-hover-color: #4b89bd;
   /* --ifm-background-color: #252628; FIXME:*/
   --ifm-background-section-dark: #3d3f42;
@@ -100,6 +102,10 @@ html[data-theme="dark"] .navbar-github-link:before {
 
 .alert--warning a {
   color: #4d3800;
+}
+
+.avatar {
+  column-gap: normal !important;
 }
 
 html[data-theme="dark"] .alert--warning, html[data-theme="dark"] .alert--warning a {
@@ -248,24 +254,6 @@ a.menu__link.menu__link--sublist {
   color: var(--ifm-heading-color);
 }
 
-/* DocItem Features */
-.zowe-faq[open] summary:before {
-  content: "📖 Click to hide answer";
-}
-
-details[open] summary:before {
-  content: "📖 ";
-}
-
-.zowe-faq summary:before {
-  content: "📘 Click to open answer";
-  margin-right: 5px;
-}
-
-summary:before {
-  content: "📘 ";
-}
-
 /* Docs Rating Box */
 .docsRating {
   display: flex;
@@ -326,7 +314,7 @@ html[data-theme="light"] .darkTheme {
   display: none;
 }
 
-.markdown h1:first-of-type {
+.markdown h1:first-child {
   display: none;
 }
