+++ src/theme/DocItem/index.js
@@ -1,143 +1,24 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useRef, useState } from "react";
 import readingTime from "reading-time";
-import Head from "@docusaurus/Head";
-import MDXComponents from "@theme/MDXComponents";
-import { MDXProvider } from "@mdx-js/react";
-import {
-  useDocsVersion,
-  useTitleFormatter,
-} from "@docusaurus/theme-common/internal";
-import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
-import useBaseUrl from "@docusaurus/useBaseUrl";
-import DocPaginator from "@theme/DocPaginator";
-import DocVersionBanner from "@theme/DocVersionBanner";
-import TOC from "@theme/TOC";
-import clsx from "clsx";
-import styles from "./styles.module.css";
-import {
-  useActivePlugin,
-  useVersions,
-} from "@docusaurus/plugin-content-docs/client";
+import { DocProvider } from "@docusaurus/theme-common/internal";
 
 //Components
-import DocsInfo from "./DocsInfo";
-import DocsRating from "./DocsRating";
+import { DocContent } from "./DocContent";
 
 function DocItem(props) {
-  const { siteConfig } = useDocusaurusContext();
-  const { url: siteUrl } = siteConfig;
-  const { content: DocContent } = props;
-  const versionMetadata = useDocsVersion();
-  const {
-    metadata,
-    frontMatter: {
-      image: metaImage,
-      keywords,
-      hide_title: hideTitle,
-      hide_table_of_contents: hideTableOfContents,
-    },
-    toc,
-  } = DocContent;
-  const {
-    description,
-    title,
-    permalink,
-    editUrl,
-    lastUpdatedAt,
-    lastUpdatedBy,
-    unversionedId,
-  } = metadata;
-  const { pluginId } = useActivePlugin({ failfast: true });
-
-  const versions = useVersions(pluginId);
-
-  const showVersionBadge = versions.length > 1;
-  const metaTitle = useTitleFormatter(title);
-  const metaImageUrl = useBaseUrl(metaImage, {
-    absolute: true,
-  });
-
+  const contentRef = useRef();
   const [readingTimeInWords, setReadingTimeInWords] = useState("");
 
   useEffect(() => {
-    setReadingTimeInWords(
-      readingTime(document.querySelector(".markdown").innerText).text
-    );
-  });
-
+    if (contentRef.current) {
+      const readTime = readingTime(contentRef.current.innerText);
+      setReadingTimeInWords(readTime.text);
+    }
+  }, [contentRef]);
   return (
-    <>
-      <Head>
-        <title>{metaTitle}</title>
-        <meta property="og:title" content={metaTitle} />
-        {description && <meta name="description" content={description} />}
-        {description && (
-          <meta property="og:description" content={description} />
-        )}
-        {keywords && keywords.length && (
-          <meta name="keywords" content={keywords.join(",")} />
-        )}
-        {metaImage && <meta property="og:image" content={metaImageUrl} />}
-        {metaImage && <meta name="twitter:image" content={metaImageUrl} />}
-        {metaImage && (
-          <meta name="twitter:image:alt" content={`Image for ${title}`} />
-        )}
-        {permalink && <meta property="og:url" content={siteUrl + permalink} />}
-        {permalink && <link rel="canonical" href={siteUrl + permalink} />}
-      </Head>
-
-      <div className="row">
-        <div
-          className={clsx("col", {
-            [styles.docItemCol]: !hideTableOfContents,
-          })}
-        >
-          <DocVersionBanner versionMetadata={versionMetadata} />
-          <div className={styles.docItemContainer}>
-            <article className="article-content">
-              {showVersionBadge && (
-                <div>
-                  <span className="badge badge--secondary">
-                    Version: {versionMetadata.label}
-                  </span>
-                </div>
-              )}
-              {!hideTitle && (
-                <header>
-                  <h1 className={styles.docTitle}>{title}</h1>
-                </header>
-              )}
-              {(editUrl || lastUpdatedAt || lastUpdatedBy) && (
-                <DocsInfo
-                  editUrl={editUrl}
-                  lastUpdatedAt={lastUpdatedAt}
-                  lastUpdatedBy={lastUpdatedBy}
-                  readingTimeInWords={readingTimeInWords}
-                  title={title}
-                />
-              )}
-              <MDXProvider components={MDXComponents}>
-                <div className="markdown">
-                  <DocContent />
-                </div>
-              </MDXProvider>
-            </article>
-
-            <div className="margin-left--none margin-top--md text--center">
-              <DocsRating label={unversionedId} />
-            </div>
-            <div className="margin-vert--lg">
-              <DocPaginator metadata={metadata} />
-            </div>
-          </div>
-        </div>
-        {!hideTableOfContents && toc && (
-          <div className="col col--3">
-            <TOC toc={toc} />
-          </div>
-        )}
-      </div>
-    </>
+    <DocProvider content={props.content}>
+      <DocContent Content={props.content} contentRef={contentRef} readingTimeInWords={readingTimeInWords} />
+    </DocProvider>
   );
 }
