+++ src/theme/MDXComponents.js
@@ -1,15 +1,15 @@
-import InitialMDXComponents from "@theme/MDXComponents";
+import InitialComponents from '@theme-original/MDXComponents';
 import Badge from "./Badge";
 import tpsr from "./tpsr";
 import img from "./ZoomedImage";
 import image from "./image";
 
 const MDXComponents = {
-  ...InitialMDXComponents,
-  Badge: Badge,
-  tpsr: tpsr,
-  img: img,
-  image: image,
+  ...InitialComponents,
+  Badge,
+  tpsr,
+  img,
+  image,
 };
 
 export default MDXComponents;
