+++ src/theme/tpsr.js
@@ -1,12 +1,16 @@
 import React from "react";
 import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
+import {
+  useDocsVersion,
+} from "@docusaurus/theme-common/internal";
 
 function tpsr() {
   const { siteConfig } = useDocusaurusContext();
-  const latestVersion = siteConfig.customFields.latestVersion;
+  const latestVersion = useDocsVersion();
+  const version = latestVersion?.label.split(" ");
   const tpsrLatestLink =
     "https://github.com/zowe/docs-site/blob/master/tpsr/tpsr-" +
-    latestVersion +
+    version[0] +
     ".md";
   return (
     <a href={tpsrLatestLink} target="_blank">
