+++ versionsArchived.json
@@ -1,5 +1,5 @@
 {
-  "v2..x LTS": "https://deploy-preview-2745--zowe-docs-master.netlify.app/v2.2.x/getting-started/overview",
+  "v2.2.x LTS": "https://deploy-preview-2745--zowe-docs-master.netlify.app/v2.2.x/getting-started/overview",
   "v2.1.x LTS": "https://deploy-preview-2745--zowe-docs-master.netlify.app/v2.1.x/getting-started/overview",
   "v2.0.x LTS": "https://deploy-preview-2745--zowe-docs-master.netlify.app/v2.0.x/getting-started/overview",
   "v1.27.x LTS": "https://deploy-preview-2531--zowe-docs-master.netlify.app/v1.27.x/getting-started/overview",
